refactor: convert file to esm
- convert file to esm
- adjust file accordingly
- update snapshots
